Test token related views.
