+++ utils/tests.py
@@ -13,9 +13,7 @@ class ViewTestCase(TestCase):
     def setUp(self):
         self.model = None
         self.credentials = {"username": "dummy", "password": "dummy"}
-        self.user = User.objects.create_user(**self.credentials)
-        self.user.is_superuser = True
-        self.user.save()
+        self.user = User.objects.create_user(**self.credentials, is_superuser=True)
 
     def authenticate_user(self):
         # Login
